# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/analysis/resumes/{resume_id}": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analyze Resume",
        "operationId": "analyze_resume_analysis_resumes__resume_id__post",
        "parameters": [
          {
            "name": "resume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resume Id"
            }
          },
          {
            "name": "jwt",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Jwt"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeResumeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "List Resume Analyses",
        "operationId": "list_resume_analyses_analysis_resumes__resume_id__get",
        "parameters": [
          {
            "name": "resume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resume Id"
            }
          },
          {
            "name": "jwt",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Jwt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisResponse"
                  },
                  "title": "Response List Resume Analyses Analysis Resumes  Resume Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/{analysis_id}/feedback": {
      "put": {
        "tags": [
          "analysis"
        ],
        "summary": "Submit Analysis Feedback",
        "operationId": "submit_analysis_feedback_analysis__analysis_id__feedback_put",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Analysis Id"
            }
          },
          {
            "name": "jwt",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Jwt"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisFeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisFeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisFeedbackRequest": {
        "properties": {
          "rating": {
            "type": "string",
            "enum": [
              "up",
              "down"
            ],
            "title": "Rating"
          },
          "expected_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Score"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "rating"
        ],
        "title": "AnalysisFeedbackRequest"
      },
      "AnalysisFeedbackResponse": {
        "properties": {
          "rating": {
            "type": "string",
            "enum": [
              "up",
              "down"
            ],
            "title": "Rating"
          },
          "expected_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Score"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "rating",
          "expected_score",
          "comment",
          "updated_at"
        ],
        "title": "AnalysisFeedbackResponse"
      },
      "AnalysisResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "resume_document_id": {
            "type": "integer",
            "title": "Resume Document Id"
          },
          "job_source": {
            "type": "string",
            "enum": [
              "none",
              "catalog",
              "pasted"
            ],
            "title": "Job Source"
          },
          "job_document_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Document Id"
          },
          "job_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Title"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "ats_category": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Ats Category"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "findings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Findings"
          },
          "years_of_experience": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Of Experience"
          },
          "technologies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Technologies"
          },
          "companies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Companies"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "feedback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisFeedbackResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "resume_document_id",
          "job_source",
          "job_document_id",
          "job_title",
          "score",
          "ats_category",
          "summary",
          "findings",
          "years_of_experience",
          "technologies",
          "companies",
          "created_at"
        ],
        "title": "AnalysisResponse"
      },
      "AnalyzeResumeRequest": {
        "properties": {
          "job_source": {
            "type": "string",
            "enum": [
              "none",
              "catalog"
            ],
            "title": "Job Source"
          },
          "job_document_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Document Id"
          }
        },
        "type": "object",
        "required": [
          "job_source"
        ],
        "title": "AnalyzeResumeRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}